+++ plugins/frontend/backstage-plugin-github-pull-requests/src/components/Home/RequestedReviewsCard/Content.tsx
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2021 Larder Software Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React from 'react';
+
+import { PullRequestsListView } from '../../PullRequestsListView';
+import { useGithubSearch } from '../../useGithubSearch';
+
+export const Content = () => {
+  const { loading, error, value } = useGithubSearch(
+    identity => `is:open is:pr review-requested:${identity.id} archived:false`,
+  );
+
+  if (loading) return <>Loading...</>;
+  if (error) return <>Error...</>;
+
+  if (value) {
+    return (
+      <PullRequestsListView
+        data={value.items}
+        loading={loading}
+        pageSize={5}
+        page={1}
+        projectName="Requested reviews from you"
+      />
+    );
+  }
+  return <div>PRs</div>;
+};
