+++ plugins/frontend/backstage-plugin-github-pull-requests/src/components/PullRequestsListView/PullRequestsListView.tsx
@@ -1,10 +1,23 @@
+/*
+ * Copyright 2021 Larder Software Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import React from 'react';
-import { useApi, githubAuthApiRef } from '@backstage/core-plugin-api';
-import { Octokit } from '@octokit/rest';
-import { useAsync } from 'react-use';
+
 import { Grid, Box, Link } from '@material-ui/core';
 import { Typography } from '@material-ui/core';
-import { getStatusIconType } from '../../Icons';
+import { getStatusIconType } from '../Icons';
 import { makeStyles } from '@material-ui/core/styles';
 import { BackstageTheme } from '@backstage/theme';
 
@@ -26,26 +39,7 @@ const useStyles = makeStyles<BackstageTheme>((theme: BackstageTheme) => ({
   },
 }));
 
-const usePullRequest = (queryFn: CallableFunction) => {
-  const auth = useApi(githubAuthApiRef);
-
-  return useAsync(async () => {
-    const token = await auth.getAccessToken(['repo']);
-    const identity = await auth.getBackstageIdentity();
-
-    const pullRequestResponse = await new Octokit({
-      auth: token,
-      baseUrl: 'https://api.github.com',
-    }).search.issuesAndPullRequests({
-      q: queryFn(identity),
-      per_page: 5,
-      page: 1,
-    });
-    return pullRequestResponse.data;
-  }, [auth]);
-};
-
-const PullRequestsListView = props => {
+export const PullRequestsListView = props => {
   const { data } = props;
   const classes = useStyles();
   console.log(data);
@@ -76,26 +70,4 @@ const PullRequestsListView = props => {
       ))}
     </Grid>
   );
-};
-
-export const Content = () => {
-  const { loading, error, value } = usePullRequest(
-    identity => `is:open is:pr author:${identity.id} archived:false`,
-  );
-
-  if (loading) return <>Loading...</>;
-  if (error) return <>Error...</>;
-
-  if (value) {
-    return (
-      <PullRequestsListView
-        data={value.items}
-        loading={loading}
-        pageSize={5}
-        page={1}
-        projectName="Your open pull requests"
-      />
-    );
-  }
-  return <div>PRs</div>;
 };
